requirements.txt
# web server
# https://pypi.org/project/Bootstrap-Flask/
Bootstrap-Flask
#https://pypi.org/project/Flask-Login/
Flask-Login
# https://pypi.org/project/Flask-HTTPAuth/
Flask-HTTPAuth
# https://pypi.org/project/Flask-SQLAlchemy/
Flask-SQLAlchemy
# https://pypi.org/project/SQLAlchemy-Utils/
SQLAlchemy-Utils
# https://pypi.org/project/Flask-WTF/
Flask-WTF
# requests (for forwarding)
# https://pypi.org/project/requests/
requests
# basic auth
#https://pypi.org/project/pydevd/
pydevd
